pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["marcos_code*"]
exclude = ["Examples*",]

[project]
name = "marcos-code"
version = "0.1.0"
authors = [{name = "Marcos Loaiza", email = "mloaizaa@unal.edu.co"}]
description = "My master thesis code"
readme = "README.md"
keywords = ["Medical Science Apps", "Topomap", "Plotting", "Utils"]
license = {text = "Simplified BSD License"}
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Education",
    "Intended Audience :: Healthcare Industry",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: BSD License",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.10",
    "Topic :: Scientific/Engineering",
    "Topic :: Scientific/Engineering :: Human Machine Interfaces",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
    "Topic :: Software Development :: Libraries",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "numpy",
    "matplotlib",
    "gcpds-visualizations @ git+https://git@github.com/UN-GCPDS/python-gcpds.visualizations.git"
]